requirements.txt
colorama
paramiko
python-nmap